Working proc macro for auth parse
