+++ hecate/Cargo.toml
@@ -14,6 +14,7 @@ readme = "README.md"
 categories = ["database"]
 
 [dependencies]
+hecate_derive = { path = "../hecate_derive" }
 reqwest = "0.9.20"
 tilecover = "1.1.1"
 crossbeam = "0.7"
