+++ hecate/src/auth/mod.rs
@@ -2,11 +2,11 @@ use crate::err::HecateError;
 use actix_http::httpmessage::HttpMessage;
 use actix_web::http::header::{HeaderName, HeaderValue};
 
-pub mod derive;
 pub mod config;
 pub mod middleware;
 pub use config::AuthContainer;
 pub use config::AuthModule;
+pub use config::AuthParse;
 pub use config::CustomAuth;
 pub use config::RW;
 use crate::user::token::Scope;
