+++ hecate_derive/Cargo.toml
@@ -1,4 +1,5 @@
 [package]
+publish = false
 name = "hecate_derive"
 version = "0.1.0"
 edition = "2018"
